pyproject.toml
[project]
name = "nonebot-plugin-cave"
version = "0.1.0"
description = "A plugin to "
authors = [{ name = "student_2333", email = "lgc2333@126.com" }]
dependencies = [
    "nonebot2>=2.0.0",
    "pydantic>=1.10.4,<2",
    "nonebot-plugin-alconna>=0.31.0",
]
requires-python = ">=3.8,<4.0"
readme = "README.md"
license = { text = "MIT" }

[project.urls]
homepage = "https://github.com/hmzz804/nonebot_plugin_cave"

[tool.pdm.build]
includes = []

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
